+++ riverpie/example/lib/extension.dart
@@ -7,7 +7,7 @@ final counter = StateProvider((ref) => 0);
 void main() {
   runApp(
     RiverpieScope(
-      observer: kDebugMode ? const RiverpieDebugObserver() : null,
+      observer: kDebugMode ? RiverpieDebugObserver() : null,
       child: const MyApp(),
     ),
   );