+++ riverpie/example/lib/view_model.dart
@@ -52,7 +52,7 @@ class Counter extends Notifier<int> {
 void main() {
   runApp(
     RiverpieScope(
-      observer: kDebugMode ? const RiverpieDebugObserver() : null,
+      observer: kDebugMode ? RiverpieDebugObserver() : null,
       child: const MyApp(),
     ),
   );