+++ riverpie/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.20.0
 
 - feat: add error handling for redux actions
+- feat: access `ref` within observers
 - feat: improve `RiverpieTracingPage` UI
 
 ## 0.19.0